QA/llm/requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
requests
langchain
langchain-community
langchain_ollama
langchain_experimental
pymysql
python-multipart
sentence_transformers
chromadb
langchain-chroma
langchain-huggingface>=0.0.2
langchain-chroma>=0.0.2
bcrypt==4.0.1